pyproject.toml
[tool.poetry]
name = "geonames"
version = "0.1.0"
description = "A package for managing and querying geonames data"
authors = ["Aareon Sullivan <askully13@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.2"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
loguru = "^0.7.2"
pandas = "^2.2.2"
aiohttp = "^3.10.5"
aiofiles = "^24.1.0"
pytest-aiohttp = "^1.0.5"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"
aioresponses = "^0.7.6"
click = "^8.1.7"
aiosqlite = "^0.20.0"
tqdm = "^4.66.5"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
black = "^21.5b1"
isort = "^5.8.0"
mypy = "^0.910"
flake8 = "^3.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[tool.poetry.group.develop.dependencies]
ruff = "^0.6.5"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
ignore_missing_imports = true
strict_optional = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
geonames-cli = "geonames.cli:cli"
